refactor: move upgradeTool to its own file
